+++ src/features/product/productPost/gallery/components/SingleImage.tsx
@@ -1,0 +1,13 @@
+import type { Product } from '@typings/products/product';
+
+interface SingleImageProps {
+  product: Product;
+}
+
+export default function SingleImage({ product }: SingleImageProps) {
+  return (
+    <div className="relative overflow-hidden rounded-md">
+      <img src={product.image} />
+    </div>
+  );
+}