+++ src/features/product/productPost/gallery/components/overlay/Icons.tsx
@@ -1,13 +1,23 @@
+import type { Product } from '@typings/products/product';
+
 import ExpandIcon from './ExpandIcon';
 import CompressIcon from './CompressIcon';
 import CloseIcon from './CloseIcon';
 
-export default function Icons({ overlayExpand, selectedOverlayIndex, handleOverlayActiveChange, handleOverlayExpandChange, singleProduct }) {
+interface IconsProps {
+  product: Product;
+  overlayExpand: boolean;
+  selectedOverlayIndex: number;
+  handleOverlayActiveChange: () => void;
+  handleOverlayExpandChange: () => void;
+}
+
+export default function Icons({ product, overlayExpand, selectedOverlayIndex, handleOverlayActiveChange, handleOverlayExpandChange }: IconsProps) {
   return (
     <div className="flex justify-between px-2.5 pt-2.5">
-      {singleProduct.gallery ? (
-        <span className="text-white" aria-label={`Image ${selectedOverlayIndex + 1} of ${singleProduct.gallery.length}`}>
-          {selectedOverlayIndex + 1}/{singleProduct.gallery.length}
+      {product.gallery ? (
+        <span className="text-white" aria-label={`Image ${selectedOverlayIndex + 1} of ${product.gallery.length}`}>
+          {selectedOverlayIndex + 1}/{product.gallery.length}
         </span>
       ) : (
         <p className="text-white" aria-label="Image 1 of 1">